+++ app/components/splineScene.tsx
@@ -1,0 +1,10 @@
+import Spline from "@splinetool/react-spline/next";
+
+const SplineScene = () => {
+  return (
+    <div>
+      <Spline scene="https://prod.spline.design/6RCYMymMCCvnA1S2/scene.splinecode" />
+    </div>
+  );
+};
+export default SplineScene;